# specs/002-auth-user-management/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Todo App Authentication API
  description: |
    Authentication and user management API for the multi-user todo application.
    Implements user registration, login, and JWT token-based authentication.
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
servers:
  - url: http://localhost:8000
    description: Local development server
  - url: https://api.example.com
    description: Production server
...
paths:
  /api/auth/register:
    post:
      tags:
        - Authentication
      summary: Register new user
      description: |
        Create a new user account with email and password.
        Email must be unique and follow RFC 5322 format.
        Password must meet complexity requirements.
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
            examples:
              valid:
                summary: Valid registration
                value:
                  email: user@example.com
                  password: SecurePass123!
      responses:
        '201':
          description: User successfully registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterResponse'
              examples:
                success:
                  value:
                    user_id: 550e8400-e29b-41d4-a716-446655440000
                    email: user@example.com
                    message: User registered successfully
        '400':
          description: Invalid input (validation error)
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              examples:
                invalidEmail:
                  summary: Invalid email format
                  value:
                    type: invalid-email-format
                    title: Invalid Email Format
                    status: 400
                    detail: Email address must be in valid format (local@domain)
                    instance: /api/auth/register
                invalidPassword:
                  summary: Invalid password
                  value:
                    type: invalid-password
                    title: Invalid Password
                    status: 400
                    detail: Password must be 8-128 characters with uppercase, lowercase, number, and special character
                    instance: /api/auth/register
                emailTooLong:
                  summary: Email too long
                  value:
                    type: validation-error
                    title: Validation Error
                    status: 400
                    detail: Email must not exceed 254 characters
                    instance: /api/auth/register
        '409':
          description: Email already registered
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: email-already-registered
                title: Email Already Registered
                status: 409
                detail: An account with this email address already exists
                instance: /api/auth/register
  /api/auth/login:
    post:
      tags:
        - Authentication
      summary: User login
      description: |
        Authenticate user with email and password.
        Returns JWT token on success.
        Rate limited to 5 attempts per minute per IP address.
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            examples:
              valid:
                summary: Valid login
                value:
                  email: user@example.com
                  password: SecurePass123!
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
              example:
                access_token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                token_type: bearer
                expires_in: 86400
        '401':
          description: Invalid credentials
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: invalid-credentials
                title: Invalid Credentials
                status: 401
                detail: Invalid email or password
                instance: /api/auth/login
        '429':
          description: Rate limit exceeded
          headers:
            Retry-After:
              description: Seconds to wait before retrying
              schema:
                type: integer
              example: 60
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: rate-limit-exceeded
                title: Too Many Requests
                status: 429
                detail: Too many login attempts. Please try again in 60 seconds
                instance: /api/auth/login
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token obtained from /api/auth/login endpoint.
        Include in Authorization header as: Bearer <token>
  schemas:
    RegisterRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          maxLength: 254
          description: User email address (RFC 5322 format)
          example: user@example.com
        password:
          type: string
          format: password
          minLength: 8
          maxLength: 128
          description: |
            Password must contain:
            - At least 8 characters
            - At most 128 characters
            - At least one uppercase letter
            - At least one lowercase letter
            - At least one number
            - At least one special character
          example: SecurePass123!
    RegisterResponse:
      type: object
      required:
        - user_id
        - email
        - message
      properties:
        user_id:
          type: string
          format: uuid
          description: Unique user identifier
          example: 550e8400-e29b-41d4-a716-446655440000
        email:
          type: string
          format: email
          description: Registered email address (lowercase)
          example: user@example.com
        message:
          type: string
          description: Success message
          example: User registered successfully
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          description: User email address
          example: user@example.com
        password:
          type: string
          format: password
          description: User password
          example: SecurePass123!
    LoginResponse:
      type: object
      required:
        - access_token
        - token_type
        - expires_in
      properties:
        access_token:
          type: string
          description: JWT access token
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJ0b2RvLWFwcC1iYWNrZW5kIiwic3ViIjoiNTUwZTg0MDAtZTI5Yi00MWQ0LWE3MTYtNDQ2NjU1NDQwMDAwIiwiZXhwIjoxNzA2MDI1NjAwLCJpYXQiOjE3MDU5MzkxMDAsInVzZXJfaWQiOiI1NTBlODQwMC1lMjliLTQxZDQtYTcxNi00NDY2NTU0NDAwMDAiLCJlbWFpbCI6InVzZXJAZXhhbXBsZS5jb20ifQ.signature
        token_type:
          type: string
          enum: [bearer]
          description: Token type (always "bearer")
          example: bearer
        expires_in:
          type: integer
          description: Token expiration time in seconds (24 hours = 86400)
          example: 86400
...
    ProblemDetails:
      type: object
      required:
        - type
        - title
        - status
        - detail
        - instance
      properties:
        type:
          type: string
          description: Error type identifier
          example: invalid-credentials
        title:
          type: string
          description: Short, human-readable error title
          example: Invalid Credentials
        status:
          type: integer
          description: HTTP status code
          example: 401
        detail:
          type: string
          description: Detailed error message
          example: Invalid email or password
        instance:
          type: string
          description: URI of the request that caused the error
          example: /api/auth/login
      description: RFC 7807 Problem Details format for error responses